Ejercicios 5, 10, 15 y 20 - Abigail Teran

+++ src/ar/edu/unju/escmi/tp2/ejercicio5/Ejercicio5.java
@@ -5,18 +5,18 @@ public class Ejercicio5 {
         int a = 8, b = 3, c = -5;
         int r1 = a + b + c;
         int r2 = 2 * b + 3 * (a - c);
-        double r3 = (double) a / b;
+        float r3 = (float) a / b;
         int r4= a % b; 
-        double r5= (double) a / c;
+        float r5= (float) a / c;
         int r6 = a % c; 
-        double r7 = (double) a * b / c;   
-        double r8 = a * ((double) b / c); 
+        float r7 = (float) a * b / c;   
+        float r8 = a * ((float) b / c); 
         int r9 = (a * c) % b;
         int aux=(c % b); if(aux<0)  aux=1;  
         int r10 = a * aux; 
         int r11 = (3 * a - 2 * b) % (2 * a - c);  
-        double r12 = 2 * ((double) a / 5 + (4 - b * 3)) % (a + c - 2);   
-        double r13 = (double) (a - 3 * b) % (c + 2 * a) / (a - c); 
+        float r12 = 2 * ((float) a / 5 + (4 - b * 3)) % (a + c - 2);   
+        float r13 = (float) (a - 3 * b) % (c + 2 * a) / (a - c); 
         int r14 = a - b - c * 2;
 
         System.out.println(r1);
